Move Input{,File}Generator out of ooni_test.hpp
Try to follow the principle that each class must be into its
own file, so it's easier to start hacking with the code.
